*桌面歌词 *部分用户点击反馈崩溃 *5.0以下系统仍可以开启导航栏变色 *修复在最近添加列表 *修复从歌曲列表添加歌曲到播放列表添加失败的问题 *蓝牙耳机拔出后自动停止播放 *修复从通知栏进入播放界面后点击关闭按钮动画异常的问题 *修复部分机型打开播放队列可能崩溃的问题

+++ app/src/main/java/remix/myplayer/ui/dialog/AddtoPlayListDialog.java
@@ -67,7 +67,7 @@ public class AddtoPlayListDialog extends BaseDialogActivity implements LoaderMan
         setContentView(R.layout.dialog_addto_playlist);
         ButterKnife.bind(this);
 
-        mAudioID = (int)getIntent().getExtras().getLong("Id");
+        mAudioID = getIntent().getExtras().getInt("Id");
 
         mAdapter = new AddtoPlayListAdapter(this);
         mAdapter.setOnItemClickLitener(new OnItemClickListener() {
